% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezdev-rstudio-plugins.R
\name{docHeader}
\alias{docHeader}
\title{Insert document header}
\usage{
docHeader()
}
\description{
Insert document skeleton header. Extracts information from
current file, R Studio project, and package structures to populate header.
}

