% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezdev-rstudio-plugins.R
\name{funcHeader}
\alias{funcHeader}
\title{Insert function header}
\usage{
funcHeader()
}
\description{
With an R script open in R Studio, place the cursor on a
function you have written and call this function - it will insert a
roxygen2 skeleton function header template.
}

